requirement.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==5.1.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
Pillow==8.1.2
prompt-toolkit==3.0.19
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
suds==0.3.5
suds-py3==1.4.4.1
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
